requirements.txt
asgiref==3.3.4
Django==4.0.4 # це джанго карл, джанго
django-ckeditor==6.1.0 # docs
django-cleanup==5.2.0 # автоматический удаляет файлы для ПолеФайлов и ПолеИзображений, после сохранения
django-js-asset==1.2.2 # Статические файлы форм
djangorestframework==3.12.4 # framework тима, фреймворк
gunicorn==20.1.0 # WSGI сервер для UNIX написанный на Python. У него нет сторонних зависимостей, легко установить и использовать.
Pillow==8.2.0 # Библиотека для изображений
psycopg2-binary==2.9.1 # PostgreSQL
pytz==2021.1
sqlparse==0.4.1


